Allow grace period before hiding past events
Events are hidden as soon as they terminate, this means that, if a
meeting is running late, the event is no longer visible and it is
impossible to check when was the scheduled end time.

Add the possibility to specify a grace period in minutes, if the event
terminated less than N minutes ago, don't hide it.